Introduce GitHub Actions Test workflow
This change introduces the basic skeleton of the Test workflow, which is
meant to be run on each pull request and push. It currently covers the
basics: building the crate in various configurations, including with
minimum versions of transitive dependencies, testing it and uploading
coverage information to Codecov, as well as running clippy, rustfmt, and
rustdoc.